api-gateway/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
python-multipart==0.0.9
prometheus-fastapi-instrumentator==6.1.0
python-logging-loki==0.3.1